Rework main for better debugging; up debounce period by 3x
Peviously, main would not expose symbols for the usb class. This adds
some unsafe ;_; which will force the symbol to be type-able in gdb.
The prior symbols would have {{impl}} as part of their name (path?)
which you can't type in a gdb console.

The o key was bouncing, so I looked at a statemap and determined that
50ms (value of 100) was about 2x as long as it needed to be. I don't
notice a difference at 50ms (yay QickDraw!) so I'm keeping it. No
bounces yet.