Rework trigger signal
So I would occationally see signals like:
                      ___________  _______
Signal  _____________/           \/       \__________
                      ___________        __________
Trigger _____________/           \______/          \_

Which is actually a bug, and causes some issues when typing,
especially when using frequently used keys, such as those on
the home row.

Ideally, this would be:
                      ___________        _
Signal  _____________/           \______/ \__________
                      _____________________________
Trigger _____________/                             \_

Which is what this commit implements.

I have not renamed this Struct, as the trigger is still Quick
to Draw, but it's slow to "put guns away" (not sure about this
 analogy :/) or maybe slow to put the "paint brushes away" :D

This also updates the comments!